Fix reset CI failures
